+++ src/DeviceManager.Api/ActionFilters/TenantHeaderOperationFilter.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using DeviceManager.Api.Constants;
 using Swashbuckle.AspNetCore.Swagger;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
@@ -24,13 +25,13 @@ namespace DeviceManager.Api.ActionFilters
 
             operation.Parameters.Add(new NonBodyParameter
             {
-                Name = "tenantid",
-                In = "header",
-                Description = "tenantid",
+                Name = DefaultConstants.TenantId,
+                Default = "b0ed668d-7ef2-4a23-a333-94ad278f45d7",
+                In = DefaultConstants.Header,
+                Description = DefaultConstants.TenantIdSwaggerDescription,
                 Required = true,
-                Type = "string",
+                Type = DefaultConstants.StringInText
             });
-
         }
     }
 }