+++ src/DeviceManager.Api/Configuration/ConfigurationOptions.cs
@@ -1,12 +1,12 @@
 ﻿using DeviceManager.Api.Configuration.Settings;
-using DeviceManager.Api.Helpers;
+using DeviceManager.Api.Constants;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace DeviceManager.Api.Configuration
 {
     /// <summary>
-    /// 
+    /// Registers all the settings from configuration file in the IOC container
     /// </summary>
     public static class ConfigurationOptions
     {
@@ -17,7 +17,8 @@ namespace DeviceManager.Api.Configuration
         /// <param name="configuration">The configuration.</param>
         public static void ConfigureService(IServiceCollection services, IConfigurationRoot configuration)
         {
-            services.Configure<ConnectionSettings>(configuration.GetSection(Constants.ConnectionStrings));
+            services.Configure<ConnectionSettings>(configuration.GetSection(DefaultConstants.ConnectionStrings));
+            services.Configure<AppSettings>(configuration.GetSection(DefaultConstants.AppSettings));
         }
     }
 }