+++ src/DeviceManager.Api/Mappings/MapProfile.cs
@@ -19,8 +19,29 @@ namespace DeviceManager.Api.Mappings
             // Device ViewModel To Device
             this.CreateMap<DeviceViewModel, Device>()
                 .ForMember(dest => dest.DeviceTitle, opt => opt.MapFrom(src => src.Title))
+                .ForMember(dest => dest.DeviceCode, opt => opt.MapFrom(src => src.DeviceCode))
                 .ForMember(dest => dest.DeviceId, opt => opt.MapFrom(src => Guid.NewGuid()))
-                ;
+                .ForMember(dest => dest.DeviceGroup, opt => opt.MapFrom(src => src));
+
+            // Device ViewModel to DeviceDetail
+            this.CreateMap<DeviceViewModel, DeviceGroup>()
+                .ForMember(dest => dest.DeviceGroupId, opt => opt.MapFrom(src => Guid.NewGuid()))
+                .ForMember(dest => dest.Company, opt => opt.MapFrom(src => src.Company))
+                .ForMember(dest => dest.OperatingSystem, opt => opt.MapFrom(src => src.OperatingSystem));
+
+            // Maps Device to Device ViewModel
+            this.CreateMap<Device, DeviceViewModel>()
+                .ForMember(dest => dest.Id, opt => opt.MapFrom(src => src.DeviceId))
+                .ForMember(dest => dest.Title, opt => opt.MapFrom(src => src.DeviceTitle))
+                .ForMember(dest => dest.DeviceCode, opt => opt.MapFrom(src => src.DeviceCode))
+                .AfterMap((src, dest, context) => context.Mapper.Map(src.DeviceGroup, dest));
+
+
+            // Maps Device Group to Device ViewModel
+            this.CreateMap<DeviceGroup, DeviceViewModel>()
+                .ForMember(dest => dest.Company, opt => opt.MapFrom(src => src.Company))
+                .ForMember(dest => dest.OperatingSystem, opt => opt.MapFrom(src => src.OperatingSystem));
+
         }
     }
 }