+++ src/DeviceManager.Api/Resources/SharedResource.cs
@@ -1,0 +1,13 @@
+﻿namespace DeviceManager.Api.Resources
+{
+    /// <summary>
+    /// Shared or common resource file for whole application
+    /// </summary>
+    public class SharedResource
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public const string ShouldNotBeEmpty = nameof(ShouldNotBeEmpty);
+    }
+}